pyproject.toml
[tool.poetry]
name = "seleniumproject"
version = "0.1.0"
description = ""
authors = ["yaroslav.bobrovnyk <y.bbobrovnik@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
selenium = "^4.8.2"
pytest = "^7.2.1"
allure-pytest = "^2.12.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
